added backend verification and error interceptor

+++ chronicleDB-admin/backend/chroniclebackend.py
@@ -8,6 +8,7 @@ import requests
 import jwt, json
 import helper
 import usermanagement as um
+import validation
 
 app = Flask(__name__)
 api = Api(app)
@@ -22,8 +23,16 @@ SECRET = "secret"
 @app.route('/show_right_flank/<stream_id>', methods=['GET'])
 def showRightFlank(stream_id):
     print("Stream ID: " + stream_id)
-    print("Received request for 'show_right_flank', printing headers:\n" + str(request.headers))
-    # TODO: Verify web token against the password
+    print("Received request for 'show_right_flank'")
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.canUserRead(token, stream_id)): 
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     print("Trying to get right flank from ChronicleDB at localhost:8000/show_right_flank/" + stream_id + " ... ")
     response = requests.get("http://127.0.0.1:8000/show_right_flank/" + stream_id)
     print("Response from ChronicleDB: "+ str(response))
@@ -32,8 +41,16 @@ def showRightFlank(stream_id):
 @app.route('/stream_info/<stream_id>', methods=['GET'])
 def streamInfo(stream_id):
     print("Stream ID: " + stream_id)
-    print("Received request for 'stream_info', printing headers:\n" + str(request.headers))
-    # TODO: Verify web token against the password
+    print("Received request for 'stream_info':")
+
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.canUserRead(token, stream_id)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     print("Trying to get stream info from ChronicleDB at localhost:8000/stream_info/" + stream_id + " ... ")
     response = requests.get("http://127.0.0.1:8000/stream_info/" + stream_id)
     print("Response from ChronicleDB: "+ str(response))
@@ -41,8 +58,17 @@ def streamInfo(stream_id):
 
 @app.route('/create_stream', methods=['POST'])
 def createStream():
-    print("Received request for 'create_stream', printing headers:\n" + str(request.headers))
+    print("Received request for 'create_stream':")
     print(request.data)
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.canUserCreateStreams(token)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response = requests.post("http://127.0.0.1:8000/create_stream", request.data)
     print("Response from ChronicleDB: " + str(response))
     return {"response from ChronicleDB": str(response) } 
@@ -50,8 +76,16 @@ def createStream():
 @app.route('/insert_ordered/<stream_id>', methods=['POST'])
 def insertOrdered(stream_id):
     print("Stream ID: " + stream_id)
-    print("Received request for 'insert_ordered', printing headers:\n" + str(request.headers))
-    # TODO: Verify web token against the password
+    print("Received request for 'insert_ordered':")
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.canUserWrite(token, stream_id)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response = requests.post("http://127.0.0.1:8000/insert_ordered/" + stream_id, request.data)
     print("Response from ChronicleDB: "+ str(response))
     return {}
@@ -59,48 +93,98 @@ def insertOrdered(stream_id):
 @app.route('/query_time_travel/<stream_id>', methods=['POST'])
 def queryTimeTravel(stream_id):
     print("Stream ID: " + stream_id)
-    print("Received request for 'query_time_travel', printing headers:\n" + str(request.headers))
-    # TODO: Verify web token against the password
+    print("Received request for 'query_time_travel'")
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.canUserRead(token, stream_id)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response = requests.post("http://127.0.0.1:8000/query_time_travel/" + stream_id, request.data)
     print("Response from ChronicleDB: "+ str(response))
     return jsonify(response.json())
 
 @app.route('/show_streams', methods=['GET'])
 def showStreams():
-    print("Received request for 'show_streams', printing headers:\n" + str(request.headers))
-    # TODO: Verify web token against the password
+    print("Received request for 'show_streams'")
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+    # if not validation.isUserAdmin(request.headers["Authorization"]):
+    #     return make_response({"Access" : "denied!!"}, 403)
+
     response = requests.get("http://127.0.0.1:8000/show_streams", request.data)
-    print("Response from ChronicleDB: "+ str(response))
+    print("Response from ChronicleDB: "+ str(response.text))
     return jsonify(response.json())
 
 @app.route('/max_key/<stream_id>', methods=['GET'])
 def maxKey(stream_id):
-    print("Received request for 'max_key', printing headers:\n" + str(request.headers))
-    # TODO: Verify web token against the password
+    print("Received request for 'max_key'")
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.canUserRead(token, stream_id)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response = requests.get("http://127.0.0.1:8000/max_key/" + stream_id, request.data)
     print("Response from ChronicleDB: "+ str(response))
     return response.text
 
 @app.route('/min_key/<stream_id>', methods=['GET'])
 def minKey(stream_id):
-    print("Received request for 'min_key', printing headers:\n" + str(request.headers))
-    # TODO: Verify web token against the password
+    print("Received request for 'min_key'")
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.canUserRead(token, stream_id)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response = requests.get("http://127.0.0.1:8000/min_key/" + stream_id, request.data)
     print("Response from ChronicleDB: "+ str(response))
     return response.text
 
 @app.route('/tree_height/<stream_id>', methods=['GET'])
 def treeHeight(stream_id):
-    print("Received request for 'tree_height', printing headers:\n" + str(request.headers))
-    # TODO: Verify web token against the password
+    print("Received request for 'tree_height'")
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.canUserRead(token, stream_id)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response = requests.get("http://127.0.0.1:8000/tree_height/" + stream_id, request.data)
     print("Response from ChronicleDB: "+ str(response))
     return response.text
 
 @app.route('/system_info', methods=['GET'])
 def systemInfo():
-    print("Received request for 'system_info', printing headers:\n" + str(request.headers))
-    # TODO: Verify web token against the password
+    print("Received request for 'system_info'")
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response = requests.get("http://127.0.0.1:8000/system_info", request.data)
     print("Response from ChronicleDB: "+ str(response))
     return response.text
@@ -126,7 +210,15 @@ def logMeIn():
 
 @app.route('/user', methods=['GET'])
 def getUsers():
-    if not um.JWTverifyToken("dummy_token"): return make_response({"Access" : "denied!!"}, 403)
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.isUserAdmin(token)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     print("User IDs request received. Printing request:\n" + str(request.data))
     res_body = um.getUserIds()
     response = make_response({"ids" : res_body}, 200)
@@ -136,7 +228,15 @@ def getUsers():
 
 @app.route('/user/<user_id>', methods=['GET'])
 def getInfoById(user_id):
-    if not um.JWTverifyToken("dummy_token"): return make_response({"Access" : "denied!!"}, 403)
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.isUserAdmin(token)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     print("User " + user_id + " request received. Printing request:\n" + str(request.data))
     res_body = um.getUserAsMap(user_id)
     response = make_response({"ids" : res_body}, 200)
@@ -146,10 +246,28 @@ def getInfoById(user_id):
 
 @app.route('/allusers', methods=['GET'])
 def getAllUsers():
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.isUserAdmin(token)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     return make_response(um.getAllUsers(), 200)
 
 @app.route('/create_user',methods=['POST'])
 def createUser():
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.isUserAdmin(token)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response=json.loads(request.data)
     um.registerNewUser(response["username"],response["password"],response["isAdmin"],response["canCreateStream"],
     response["allowedStreams"],response["allowedInsertStreams"],response["allStreamsAllowed"],response["canInsertAll"])
@@ -158,12 +276,30 @@ def createUser():
 
 @app.route('/delete_user', methods=['POST'])
 def deleteUser():
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.isUserAdmin(token)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response=json.loads(request.data)
     um.deleteUser(response)
     return{}
 
 @app.route('/exists_user', methods =['POST'])
 def existsUser():
+    
+    # Validation #####################################
+    print("Validating Token:")
+    token = request.headers["Authorization"]
+    if (token): print("Found Header")
+    if not (validation.verifyJWT(token) and validation.isUserAdmin(token)):  
+        return make_response({"Access" : "denied!!"}, 403)
+    # Method #########################################
+
     response=json.loads(request.data)
     tmp=um.userAlreadyExists(response)
     return make_response(jsonify(tmp),200)
